requirements.txt
# Core
python>=3.9

# Deep Learning
torch>=2.0
transformers>=4.36
huggingface-hub>=0.19

# Data & ML
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2

# Utilities
pyyaml>=6.0
tqdm>=4.64

# Optional (safe to keep)
sentencepiece>=0.1.99